tags-test-suites
We cannot put Before and After code into separate class because of this error:
io.cucumber.java.InvalidMethodException: You're not allowed to extend classes that define Step Definitions or hooks. class stepDefinitions.contactUsSteps extends class stepDefinitions.base.Hooks
