+++ src/test/java/stepDefinitions/base/Hooks.java
@@ -1,0 +1,27 @@
+package stepDefinitions.base;
+
+import io.cucumber.java.After;
+import io.cucumber.java.Before;
+import org.openqa.selenium.PageLoadStrategy;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+
+public class Hooks {
+
+    public WebDriver driver;
+
+    @Before
+    public void setup_Chrome() {
+        System.setProperty("webdriver.chrome.driver", System.getProperty("user.dir") + "/src/main/java/drivers/chromedriver.exe");
+        ChromeOptions chromeOptions = new ChromeOptions();
+        chromeOptions.setPageLoadStrategy(PageLoadStrategy.NORMAL);
+        driver = new ChromeDriver(chromeOptions);
+        driver.manage().window().maximize();
+    }
+
+    @After
+    public void tearDown() {
+        driver.quit();
+    }
+}
